+++ prp/parse/metadata.py
@@ -1,7 +1,7 @@
 """Parse metadata passed to pipeline."""
 import json
 import logging
-from typing import List, TextIO
+from typing import List
 
 from Bio import SeqIO
 
@@ -10,40 +10,41 @@ from ..models.metadata import RunInformation, SoupVersion
 LOG = logging.getLogger(__name__)
 
 
-def get_database_info(process_metadata: List[TextIO]) -> List[SoupVersion]:
+def get_database_info(process_metadata: List[str]) -> List[SoupVersion]:
     """Get database or software information.
 
     :param process_metadata: List of file objects for db records.
-    :type process_metadata: List[TextIO]
+    :type process_metadata: List[str]
     :return: Description of software or database version.
     :rtype: List[SoupVersion]
     """
     db_info = []
-    for soup in process_metadata:
-        dbs = json.load(soup)
-        if isinstance(dbs, (list, tuple)):
-            for db in dbs:
-                db_info.append(SoupVersion(**db))
-        else:
-            db_info.append(SoupVersion(**dbs))
+    for soup_filepath in process_metadata:
+        with open(soup_filepath, "r", encoding="utf-8") as soup:
+            dbs = json.load(soup)
+            if isinstance(dbs, (list, tuple)):
+                for db in dbs:
+                    db_info.append(SoupVersion(**db))
+            else:
+                db_info.append(SoupVersion(**dbs))
     return db_info
 
 
-def parse_run_info(run_metadata: TextIO) -> RunInformation:
+def parse_run_info(run_metadata: str) -> RunInformation:
     """Parse nextflow analysis information
 
     :param run_metadata: Nextflow analysis metadata in json format.
-    :type run_metadata: TextIO
+    :type run_metadata: str
     :return: Analysis metadata record.
     :rtype: RunMetadata
     """
     LOG.info("Parse run metadata.")
-    run_info = RunInformation(**json.load(run_metadata))
+    with open(run_metadata, "r") as jsonfile:
+        run_info = RunInformation(**json.load(jsonfile))
     return run_info
 
 
-def get_gb_genome_version(gff_path: str) -> str:
+def get_gb_genome_version(fasta_path: str) -> str:
     """Retrieve genbank genome version"""
-    handler = SeqIO.parse(gff_path, "gb")
-    record = next(handler.records)
-    return record.id, record.description
+    record = next(SeqIO.parse(fasta_path, "fasta"))
+    return record.id, record.description.rstrip(", complete genome")