+++ prp/parse/species.py
@@ -6,7 +6,7 @@ import pandas as pd
 LOG = logging.getLogger(__name__)
 
 
-def parse_kraken_result(file: str):
+def parse_kraken_result(file: str, cutoff: float = 0.0001):
     """parse_species_pred""Parse species prediciton result"""
     tax_lvl_dict = {
         "P": "phylum",
@@ -22,5 +22,6 @@ def parse_kraken_result(file: str):
         .sort_values("fraction_total_reads", ascending=False)
         .rename(columns=columns)
         .replace({"taxonomy_lvl": tax_lvl_dict})
+        .loc[lambda df: df['fraction_total_reads'] >= cutoff]
     )
     return species_pred.to_dict(orient="records")